pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cognify"
version = "0.1.0"
description = "Self-driving content platform — trend discovery, multi-agent research, article generation"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "slowapi>=0.1.9",
    "PyJWT[crypto]>=2.9.0",
    "bcrypt>=4.2.0",
    "email-validator>=2.2.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "pytrends>=4.9.0",
    "asyncpraw>=7.7.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "freezegun>=1.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # Depends() in FastAPI default args is idiomatic

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
disallow_any_generics = true

[[tool.mypy.overrides]]
module = "pytrends.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "asyncpraw.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langgraph.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langchain_core.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langchain_anthropic.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
fail_under = 80
show_missing = true
